First part of making Code synthesised by CertiCoq executable as part of Linux Kernel Modules: Preparing the lazily failing stack.
As the garbage collection has to run in kernel space, we first have to get rid of the loss of control flow, induced by calls to exit and assert. In Kernel space, exit is not usable. While do_exit exist, it is only meant for kthreads. But kthreads should only be used if really necessary and usually they are not. Also, using kthreads introduces both in memory and time.
Also, there is no assert in the fashion of the libc assert. While there exists some macro BUG_ON (and BUG) which could be potentially used, its use is (to express it politely, see [1]) discouraged. BUG_ON would produce a kernel panic which is not acceptable.

To solve those issues, we start with being lazy on failure, i.e. transforming the GC in a way that errors are propagated to the public functions. This way, the errors can be handled by the calling code.

[1] https://lkml.org/lkml/2016/10/4/1