First part of making Code synthesised by CertiCoq executable as part of Linux Kernel Modules: Preparing the lazily failing stack.
Step 1:
When calling the public function make_tinfo, the private functions create_heap and create_space are involved in the call graph. Here, make_tinfo first calls create_heap (which may fail), then allocates the thread_info and calls exit, if the allocation failed.
The function create_heap itself tries to allocate the heap memory, calls exit on failure. If the memory allocation succeeded, create_space is called.
Then, create_space attempts to allocate the next generation and calls exit on failure.
In order to transform this into a lazy failing approach, we start with create_space which is the bottom of the call graph. Originally, create_space is a void function, i.e. returns nothing. To catch failures, we redefine the function to be a boolean function. Then, instead of exiting on failing allocation, we return false (0) in this case and return true (1) if the memory allocation succeeded.
Now that create_space propagates failures to create_heap, we adopt the behaviour of create_heap itself.
If the memory allocation fails (i.e. we get a NULL pointer), we merely return it. Additionally, if the call
to create_space evaluates to false (0), we return NULL, too. Thus, create_heap returns NULL if and only if malloc failed or create_space failed.
This transformation allows make_tinfo to detect a failing call to create_heap. Thus, we catch a NULL value from create_heap and return NULL in this case. Also, instead of calling exit on failing allocation of a thread_info, we return NULL here, too.
Here again, make_tinfo will return NULL if and only if create_heap returned NULL or allocating tinfo failed (i.e. returned NULL)