Preparing the lazily failing stack based GC, Step 3:
This is an easy one.
The function certicoq_modify has a quite simple structure since it only sets a mutable
write barrier, does not call other functions and only fails with an assertion if
ti->alloc < ti->limit does not hold, i.e. when the alloc pointer reached or ex-
ceeded the limit. This can be solved quite easily by making the function return
booleans and, introducing a compile barrier that returns false if the asserted con-
dition is not met for Kernel space code and uses the function assert for user space
code. Also, in both cases the function return true at the end of the function body.