Preparing the lazily failing stack based GC, Step 4:
When calling the public function export_heap, the private functions create_space,
create_heap, do_generation and garbage_collect_all are involved in the call
graph. Here, the functions create_space, create_heap and do_generation were already
transformed previously.
And in fact, neither export_heap nor garbage_collect_all do fail by calling exit or assertions.
But garbage_collect_all calls create_heap and do_generation. Thus, the function
has to be transformed to propagate failures from those functions. Both functions
return false on error after their transformation. And garbage_collect_all shall
return an integer which makes the propagation more complicated but not impossible.
The solution here is that garbage_collect_all returns a value i that is
defined by an iteration in the interval [0, MAX_SPACES − 2]. Thus, every
value outside of that interval is an error value. And since the return value has to
be of type int (which is in fact signed int), a valid error return value is −1. So,
the function can be redefined to return −1 iff create_heap or do_generation fail.
Then, it is possible to catch a failing garbage_collect_all in export_heap
and the function also has to catch failing calls to create_space and do_generation.
Additionally, there are two memory allocations that create memory for spaces to
do the export. Those memory allocation calls are not guarded in any way. Thus,
a failing allocation would produce a null pointer, which needs to be handled. To
summarise, export_heap has to return NULL iff one of the memory allocations or
calls to the described functions failed. Also, it has to be made sure that successfully
allocated memory is freed when one of the functions fails.