Preparing the lazily failing stack based GC, Step 5:
After having transformed all public functions and functions called by them to
lazy failing variants, there are still private functions that have to be addressed.
One of the functions is uintnat gensize(uintnat words) which is actually even
deactivated by guarding it by a “if false” macro. This function computes the
size of the next generation and calls exit via abort_with if the size exceeds the
size of an unsigned integer, i.e. would be too big for the address space. Also, the
function asserts that the size of the next generation is at least two times as big as
the current one. Assuming that the current generation cannot have a size of 0, the
size must be positive. Thus, the size of the next generation must also be positive.
And since the return value is uintnat which includes the value 0, this value can be
used as error value. Now, the call to exit can be rewritten to returning 0. And in
the same fashion as previously, the assertion can be handled analogously as in certicoq_modify, for example.
Now that all calls to abort_with were removed, the function itself can be
removed, too. Also, no assertions are called if the code is compiled for Kernel
space. Thus, the include to the assert.h can be guarded by a compile barrier
to make sure that it is removed for Kernel space.
To make sure that the new garbage collection works as previously, it has to still
fail in user space code. Especially the call to garbage_collect from synthesised
code must still exit in user space, but not in Kernel space. And since it is the
only function of the garbage collection that is called by the synthesised code, no
other functions need to be explicitly exit as invalid values can be caught by calling
plain C code. The function garbage_collect in the implementation file can be
renamed into do_garbage_collect.
In the header file, the return type of garbage_collect is made dependent on whether the file is
compiled for Kernel or user space. In the implementation file, the new implementation
of the function garbage_collect then returns the result from
do_garbage_collect in Kernel space and exits in user space if the return value
of the function is 0, i.e. false.
With all those changes, the garbage collection is still compatible for user space
applications. While the lazy failing leads to a time latency before eventually failing,
this latency is not relevant in user space since both the original and modified
garbage collection will fail in user space and as result stop the process.