Specialising the stack based GC for Kernel space, Step 1:
The functions printf, fprintf, malloc and free cannot be used. Thus,
the specialised implementation has to use functionality provided by the Linux
Kernel headers.
The most easily solvable aspect is the printing functionality. In Kernel space, all
messages are written into the Kernel logs when using the function printk(log_level fmt_string,
from linux/printk.h. The printing function expects as input a log level, usually KERN_INFO,
a format string and if the format string includes placeholders,
additionally a colon separated list of arguments. So the main difference between
printf and printk is the log level and the different header file. Thus, all calls to
printf and fprintf can be replaced by calls to printk. Specifically, all informational
printing instructions are set to log level KERN_INFO and all error messages,
e.g. the ones introduced in the transformation into a lazily failing GC, are set to
log level KERN_ERR. It is important to note that using printk to write into files
is discouraged since writing into files that are not located in the proc directory,
is unsafe. (I have to search the source of this information again)
Thus, the function printtree effectively cannot be used and the function in_heap
will never be called. While these functions are only
compiled if the flag DEBUG is set, for Kernel space applications those functions should also
be ignored. This is done by compiling them only if CERTICOQ_KERNEL_SPACE is not defined.
Naturally, it would also be possible to just remove the functions.
Additionally, the dependant include of assert.h and all assertions and user space specific
implementations can be removed.