Specialising the stack based GC for Kernel space, Step 2:
Also, Kernel modules are compiled with many warnings enabled, like -Werror=strict-prototype
The warning, for instance, is emitted if a function declaration or definition does
not contain argument types. If a function does not have arguments, as it holds
for create_heap, the argument type of the function has to be void. And in this
case, this warning is even treated as error. It is naturally possible to override
this warning, but from an engineering perspective, code of the garbage collection
should fulfill the same constraints as any other Kernel module component does.

Also, remove annotations.