Specialising the stack based GC for Kernel space, Step 3:
What remains is the adoption of memory allocations to the Linux Kernel.

While user space code can just use malloc and free for memory allocation and
freeing heap memory, as the current GC does, allocating memory in Kernel space
is more complicated. The garbage collection allocates memory in the functions
create_heap, create_space, make_tinfo and export_heap.
It is important to note that create_space in fact allocates an array with n cells.
Thus, it is possible to use kmalloc_array in create_space and kzalloc in the
remaining functions. One hurdle is that the Kernel specific functions are
defined in linux/slab.h. And this header file transitively imports
the header thread_info.h which includes a definition of a - as can be expected
- thread_info. Thus, when including linux/slab.h instead of stdlib.h in the
garbage collection, the compilation of the garbage collection will eventually fail
due to the redefinition. But this problem can be solved by defining a new
implementation file for memory allocations (kmod_alloc.c).
Here, the include of slab.h is done in the C file, which is compiled
as separate compilation unit. And in this file, no redefinition of thread_info
will occur. In fact, the thread_info definition from the Linux Kernel header
is not even used and is thus eliminated. When exposing the functions alloc,
alloc_array and free in the header file (kmod_alloc.h) and including
the file in garbage collection instead of stdlib.h, the new functions can be used.
That is, the function alloc is used in create_heap, make_tinfo and export_heap
while alloc_array is used in create_space.
One side-effect of not using the stdlib.h is that the stdint.h which contains
definitions of integral types (e.g. size_t) is not included. But those definitions are
necessary for the memory allocation. This can also be solved easily: by including
linux/types.h in the kmod_alloc.h.
Small addition: While it is possible to free or reset the heap, there is no explicit function for
freeing the thread_info. Although it is possible to accomplish that manually by
first calling free_heap and then freeing the thread_info itself, it is easy to forget
freeing the heap. Thus, a public function free_tinfo was implemented that eases
this task.
Now, the garbage collection is fully adopted to the Linux Kernel. But there
are still other files necessary during compilation that have to be adopted.