Adopting values.h
The macros Tag_val, Tag_hp and Byte_u all return an unsigned char * as value.
But Bp_hp, Bp_val(Line 17) and Byte(Line 20) return a char *.
Although it is not noted specifically, defining Byte as char * and Byte_u
as unsigned char * suggests that the macros shall return different types.
This assumption is supported by the C ISO/IEC standard that states
“The three types char, signed char, and unsigned char are collectively
called the character types. The implementation shall define char to have the same
range, representation, and behavior as either signed char or unsigned char” [11,
§6.2.5.15]. And according to the GCC compiler documentation, the default representation
of char is determined by the ABI, i.e. platform specific [8, §4.4, item 5].
For the AMD64 ABI, for example, both char and signed char are represented as
signed byte [14, p. 16, fig. 3.1]. Thus, on a standard machine, the type char *
should have the same range, representation and behaviour as signed char *.
However, Linux Kernel modules on an average Ubuntu 22.04 x86_64 machine
are compiled with the option -funsigned-char which is contrary to the assumptions.
For instance, in this case Byte(x, i) and Byte_u(x, i) would have the
same type which may not be the anticipated result. Thus, when compiling for the
Linux Kernel, the expected representation of char * should be explicitly specified.
Under the assumption that char * shall have the semantics of signed char *,
the types were adopted for Kernel compilation respectively. Since these macros
seem to be unused, it would also be possible to eliminate them. However, they could
be useful in the future.

[8]  : https://gcc.gnu.org/onlinedocs/gcc-13.2.0/gcc/Characters-implementation.html
[11] : https://www.open-std.org/jtc1/sc22/wg14/www/docs/n1570.pdf
[14] : https://gitlab.com/x86-psABIs/x86-64-ABI/-/jobs/artifacts/master/raw/x86-64-ABI/abi.pdf?job=build